+++ src/compas_viewer/actions/__init__.py
@@ -15,10 +15,6 @@ def register(name: str, action_class: Callable):
     action_class : :class:`Action`
         Action class.
 
-    Returns
-    -------
-    None
-
     """
     ITEM_ACTIONS[name] = action_class
 
@@ -26,6 +22,7 @@ def register(name: str, action_class: Callable):
 def register_actions():
     register("zoom_selected", ZoomSelected)
     register("print_gl_info", GLInfo)
+    register("select_all", SelectAll)
 
 
 def get_action_cls(name: str) -> Any:
@@ -40,6 +37,7 @@ def get_action_cls(name: str) -> Any:
 from .action import Action  # noqa: F401 E402
 from .zoom_selected import ZoomSelected  # noqa: E402
 from .print_gl_info import GLInfo  # noqa: E402
+from .select_all import SelectAll  # noqa: E402
 
 register_actions()
 
@@ -47,4 +45,5 @@ ___all__ = [
     "Action",
     "ZoomSelected",
     "GLInfo",
+    "SelectAll",
 ]