+++ src/compas_viewer/actions/select_all.py
@@ -1,0 +1,10 @@
+from .action import Action
+
+
+class SelectAll(Action):
+    """Look at action."""
+
+    def pressed_action(self):
+        for obj in self.viewer.objects:
+            obj.is_selected = True
+        self.viewer.render.update()